[configtool] Use case-insensitive filter
During filtering, if you wanted to look for, for example "French"
or "Chinese", you had to exactly type "Fre..." or "Chi...".
Typing "fre.." or "chi..." did no show any result, which is not
very convenient for the user.

This patch uses `strcasestr' instead of `strstr' to find a substring
in the filtering function. Thus we can do a case-insensitive search.